feat(server): :recycle: centralize repo config
